Structural Update

+++ Models/DatabaseModels/EPRSDatabaseObjects/Core/OrderMain.cs
@@ -17,7 +17,7 @@ namespace Models.DatabaseModels.PermitIssuance.Core
         public long TotalAmount { get; set; }
         public long PermitNo { get; set; }
         public long TotalQuantity { get; set; }
-        public Decimal Tax { get; set; }
+        public long Tax { get; set; }
         public long Discount { get; set; }
         public long GrandTotal { get; set; }
